integrating active node management modules option in pool settings and implementing execution of only the active modules defined. fixing column definition in decisions table to prevent numeric overflow, fixing tests, refactoring.

+++ cloudify-widget-pool-manager/src/main/resources/sql/4.sql
@@ -9,4 +9,4 @@ create table decisions (
 );
 
 delete from nodes;
-alter table nodes add column expires int;
+alter table nodes add column expires bigint;
